fix: SQL 처리량을 초당 실행 건수로 보정compression

+++ sqladvisor/src/main/java/dbinc/sqladvisor/domain/monitoring/dto/MonitoringDtos.java
@@ -42,7 +42,7 @@ public final class MonitoringDtos {
     public record ActivityPoint(
             LocalDateTime collectedAt,
             long activeSessions,
-            long executions,
+            double executions,
             double cpu,
             long io
     ) {
